Assestment/requirements.txt
PyPDF2
python-docx
langchain
langchain-community
langchain-openai